Create separate template Write and WriteFile functions
For supporting general and language specific gitignore templates we need
the ability to easily write more than one template to a single
destination. This commit moves the existing functionality of writing to
a provided path to WriteFile. Write now takes an io.Writer argument
instead, allowing the chaining of multiple writes to the same
destination.