Rename and covert new dataset function to method
The New function didn't return a Dataset struct and also had the
additional side effect of generating code. Renaming it to GenerateCode
makes the behaviour much more obvious. It also only used data already
available in the dataset struct so converting it to a method makes it
more convenient to use.