Helper functions for getting canonical names
We allow users to refer to datasets with or without the file extension
since the name must be unique. However, internally, we do not use the
file extension so must always clean the user input to match. This commit
solidifies the concept of canonical names and provides helper functions
to clean user input rather than having the same code repeated in
multiple places.