Fetcher: Change interface to take writer instead of returning bytes
Since the most common and currently only use case for fetching a dataset
is to write it to disk there's no reason to be loading the entire thing
into memory and returning it just for the caller to write it. Instead we
can take a writer as an argument and do a buffered copy. This means we
never load the entire datset into memory and makes working with fetchers
more ergonomic.